Bug: ListAll crashes as it never pages
ListAll was not calling the API to retrieve the next page of records. After the GetNextURL method is fixed, this method caused an out of memory exception.

Additionally, the list class works better when given a proper Capacity value. Otherwise, the base implementation will attempt to allocate arrays with exponential growth.